chore: convert next.config to typescript
