+++ web/server/config/i18n/index.ts
@@ -1,0 +1,6 @@
+export * from './buildLocalizedRoutes';
+export * from './config';
+export * from './localeSubpaths';
+export * from './localizedRoutes';
+export * from './redirect';
+export * from './rewrite';
