+++ web/server/config/i18n/localeSubpaths.ts
@@ -1,0 +1,6 @@
+import { i18n } from './config';
+
+export const localeSubpaths = i18n.locales.reduce((localeSubpaths, locale) => {
+  localeSubpaths[locale] = locale;
+  return localeSubpaths;
+}, {});
