+++ web/server/config/i18n/localizedRoutes.ts
@@ -1,0 +1,4 @@
+import translatedPaths from 'infra/i18n/translatedPaths';
+import { buildLocalizedRoutes } from './buildLocalizedRoutes';
+
+export const localizedRoutes = translatedPaths.map((r) => buildLocalizedRoutes(r));
