+++ web/server/web.ts
@@ -1,4 +1,5 @@
 /* eslint-disable no-console */
+import 'dotenv/config';
 import next from 'next';
 import express from 'express';
 import { createProxyMiddleware } from 'http-proxy-middleware';
@@ -6,6 +7,7 @@ import csurf from 'csurf';
 import cookieParser from 'cookie-parser';
 import { setupLogging } from './logging';
 import { generateSiteMapXML } from './sitemap';
+import nextConfig from './next.config';
 
 const INTERNAL_BACKEND_TARGET = process.env.FBG_BACKEND_TARGET || 'http://localhost:3001';
 const dev = process.env.NODE_ENV !== 'production';
@@ -15,7 +17,7 @@ const STATIC_DIR = APP_DIR + 'public/static/';
 
 const PORT = process.env.SERVER_PORT || 3000;
 const isProdChannel = process.env.CHANNEL === 'production';
-const app = next({ dev });
+const app = next({ dev, conf: nextConfig });
 const handle = app.getRequestHandler();
 
 const csrfProtection = csurf({ cookie: true });
