+++ web/tsconfig.server.json
@@ -1,6 +1,9 @@
 {
   "compilerOptions": {
     "baseUrl": "./src/",
+    "paths": {
+      "server/*": ["../server/*"]
+    },
     "target": "es6",
     "lib": [
       "dom",
